Sep-16 structural hardening: contracts, proofs, gates (22 tickets + 2 audit rounds)
Implements docs/plans/sep-16-hardening (H16-001..022) against 9ae9547 and
processes two rounds of adversarial audit on top (docs/plans/sep-16-hardening/
AUDIT-2026-09-16.md). Contracts are recorded in ADR 0003 (D01-D16).

Engine (taskmesh-engine)
- one admission authority: class capacity + capability-pool occupancy in one
  transition; SubstrateSaturated (shed) vs SubstratePoolTimedOut (queued wait)
- ticket lifecycle: ClaimOutcome::{Ready,Pending,Terminal,Invalid}, bounded
  terminal retention (O(log n)); ReleaseOutcome (#[must_use]); claim touches the lease
- lock discipline: clock sampled before the lock, monotonic commit watermark,
  every host Arc woken/dropped outside the mutex, waker panics contained
- promotion: PROMOTION_BUDGET with continuation; budget checked before select;
  D08 in-class FIFO across the continuation gap (promotion_pending, QueuedBehind,
  capability-blocked head is the only same-class overtake)
- exact u128 accounting, sticky accounting fault, Snapshot schema v2
- memory: reservation/measurement/release as distinct facts, single-lock epochs,
  sweep reclaims only DispatchReserved
- fairness: WFQ 2^64 scale + weight-0 rejection, arithmetic DRR ring, idle-credit
  reset, cancellation leaves no debt
- src/sync.rs seam: loom/shuttle model-check the production Governor

Host (taskmesh)
- ResolvedExecutionPlan; a stack request consumes large_stack; DedicatedStackThread
  (sync) vs DedicatedStackRuntime (async)
- acquisition budget covers every wait (D09) on both the immediate and queued paths
- caller answer vs worker custody (D10): leases travel with results; terminal and
  panic replies precede owned-runtime teardown
- typed refusals: DeadlineUnsupported, LeaseReclaimed, WorkerPanicked,
  WorkerUnavailable, JobAbandoned; ExecutorCapabilities is load-bearing
  (ExecutorDeclaresFewerWorkers)

Proof surface
- tools/verification: mutation gate (43 entries, cargo + pytest runners, message
  scoped to the named test, anchors pre-checked, --require-clean in CI)
- tools/gates: inventory + required set + parity/enforcement validator; just proof
  expands to exactly the required set
- tools/qualification: receipts bound to source identity (digest, HEAD, dirtiness)
- tools/bench: allocation gate (threshold = measured 3.0, counter self-check);
  bench-iai fingerprinted baselines with tested shell
- semgrep pack async-aware with fire/clean fixtures for all 38 rules; arch checker
  scans every src file and every spelling; consumer-MSRV 1.81 fixture

Breaking: Governor::claim/release/release_stage_memory return outcome enums;
Snapshot schema 2; SubmitOptions::deadline on a class that cannot enforce it is
refused; worker failures are typed; a CpuExecutor declaring fewer workers than the
cpu gate is refused at build. External consumers are unknown.

Local evidence: just gate green (372 Rust / 206 Python), just matrix, loom 5/5,
shuttle 4/4, mutants 43/43, consumer MSRV 1.81 PASS. Receipt NOT_QUALIFIED only
for the dirty tree and the Linux-only bench-iai gate (both by design here).

+++ crates/taskmesh-engine/tests/memory_modes.rs
@@ -35,7 +35,7 @@ fn admit(g: &Governor) -> PermitId {
     }
 }
 
-fn held(g: &Governor, c: &TaskClass) -> u32 {
+fn held(g: &Governor, c: &TaskClass) -> u128 {
     g.snapshot().classes[c].memory_units_held
 }
